+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## [5.47.1](https://github.com/plivo/plivo-java/tree/v5.47.1) (2026-04-17)
 **Bug Fix - PhoneNumber Compliance API**
-- Fixed RequirementLister to not extend Getter, preventing id query parameter leak
-- Fixed PhoneNumberComplianceGetter to exclude id from query parameters
+- Fixed PhoneNumberComplianceGetter to exclude `id` from query parameters via toMap() override
 
 ## [5.47.0](https://github.com/plivo/plivo-java/tree/v5.47.0) (2026-04-08)
 **Feature - PhoneNumber Compliance API support**